get_token/requirements.txt
requests
pyjwt


